backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
python-dotenv
requests
langchain-openai
langgraph
langgraph-cli[inmem]
langchain
ipython
